themeToggle fixed and moved to breadcrumbs
